+++ src/main/java/com/example/paymentservice/web/dto/PaymentDto.java
@@ -3,6 +3,8 @@ package com.example.paymentservice.web.dto;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Pattern;
 
+import java.math.BigDecimal;
+
 public class PaymentDto {
     private Long orderId;
 
@@ -28,10 +30,10 @@ public class PaymentDto {
     @Pattern(regexp="^\\d{3}$", message = "CVV код должен состоять из 3 цифр")
     private String cvv;
 
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
 
     public PaymentDto(Long orderId, String cardNumber, String customerName,
-                      String dateCard, String cvv, Integer totalPrice) {
+                      String dateCard, String cvv, BigDecimal totalPrice) {
         this.orderId = orderId;
         this.cardNumber = cardNumber;
         this.customerName = customerName;
@@ -80,11 +82,11 @@ public class PaymentDto {
         this.cvv = cvv;
     }
 
-    public Integer getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Integer totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
